% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbourhoodMerger.R
\name{neighbourhoodMerger}
\alias{neighbourhoodMerger}
\title{neighbourhoodMerger Function}
\usage{
neighbourhoodMerger(neighbourhoods, verbose = TRUE)
}
\arguments{
\item{neighbourhoods}{list of objects of class "multiPhylo"}

\item{verbose}{Prints the function's progress. Defaults to TRUE}
}
\value{
List of "multiPhylo" objects
}
\description{
This function merges the 1-NNI, 2-NNI or x-SD neighbourhoods outputted by \code{\link{nniNeighbourhoods}}, \code{\link{two_nniNeighbourhoods}} and \code{\link{xSDneighbourhoods}} with shared trees, until only the tree islands are left.
}
\examples{
data(testTrees)
neighbourhoods <- nniNeighbourhoods(testTrees)
islands <- neighbourhoodMerger(neighbourhoods)
#write.islands(islands)

}
\references{
Maddison, D. R. (1991) \href{https://doi.org/10.1093/sysbio/40.3.315}{The discovery and importance of multiple islands of most-parsimonious trees}. \emph{Syst. Zool.}, 40:315-328
}
