% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nniIslands.R
\name{nniIslands}
\alias{nniIslands}
\title{nniIslands Function}
\usage{
nniIslands(neighbourhoods, verbose = TRUE)
}
\arguments{
\item{neighbourhoods}{list of objects of class "multiphylo"}

\item{verbose}{Prints the function's progress. Defaults to TRUE}
}
\value{
List of "multiphylo" objects
}
\description{
This function merges the 1-NNI neighbourhoods outputted by \code{\link{nniNeighbourhoods}} with shared trees, until only the NNI islands are left.
}
\examples{
data(testTrees)
neighbourhoods <- nniNeighbourhoods(testTrees)
islands <- nniIslands(neighbourhoods)
#write.islands(islands)

}
