% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nniNeighbourhoods.R
\name{nniNeighbourhoods}
\alias{nniNeighbourhoods}
\title{nniNeighbourhoods Function}
\usage{
nniNeighbourhoods(tree, verbose = TRUE)
}
\arguments{
\item{tree}{an object of class "multiphylo"}

\item{verbose}{Prints the function's progress. Defaults to TRUE}
}
\value{
List of "multiphylo" objects
}
\description{
This function extracts the 1-NNI neighbourhoods from a multiphylo object, it generates a list of multiphylo objects containing all trees shared between each tree's NNI neighbourhood and the multiphylo object.
At this time the package cannot place unresolved trees into islands.
}
\examples{
data(testTrees)
neighbourhoods <- nniNeighbourhoods(testTrees)
islands <- nniIslands(neighbourhoods)
#write.islands(islands)

}
\references{
Maddison (1991) Syst Zool 40:315-328
(\href{https://doi.org/10.1093/sysbio/40.3.315}{OUP})
}
