% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testTrees-data.r
\docType{data}
\name{testTrees}
\alias{testTrees}
\title{Example trees for islandNeighbours package}
\format{
An object of class "multiPhylo".
}
\source{
\href{https://github.com/anaserrasilva/MajorityRuleAndTreeIslands}{GitHub}
}
\usage{
data(testTrees)
}
\description{
Selection of ten phylogenetic trees from the set of most parsimonious trees recovered from Pardo et al.'s (2017) datamatrix.
}
\examples{
data(testTrees)
neighbourhoods <- nniNeighbourhoods(testTrees)
islands <- nniIslands(neighbourhoods)
#write.islands(islands)
}
\references{
Pardo et al. (2017) PNAS 114:5389-5395
(\href{https://pubmed.ncbi.nlm.nih.gov/28630337/}{PubMed})
}
\keyword{datasets}
