% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoNNIneighbourhoods.R
\name{two_nniNeighbourhoods}
\alias{two_nniNeighbourhoods}
\title{two_nniNeighbourhoods Function}
\usage{
two_nniNeighbourhoods(tree, verbose = TRUE)
}
\arguments{
\item{tree}{an object of class "multiPhylo"}

\item{verbose}{Prints the function's progress. Defaults to TRUE. If run in parallel set to FALSE.}
}
\value{
List of "multiPhylo" objects
}
\description{
This function extracts the 2-NNI neighbourhoods from a multiPhylo object, it generates a list of multiPhylo objects containing all trees shared between each tree's NNI neighbourhood and the multiPhylo object.
At this time the package cannot place unresolved trees into islands.
}
\examples{
data(twoNNItestTrees)
neighbourhoods <- nniNeighbourhoods(twoNNItestTrees)
islands <- nniIslands(neighbourhoods)
#write.islands(islands)

}
