% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testTrees-data.R
\docType{data}
\name{twoNNItestTrees}
\alias{twoNNItestTrees}
\title{Example trees for 2-NNI island extraction}
\format{
An object of class "multiPhylo".
}
\usage{
data(twoNNItestTrees)
}
\description{
Three phylogenetic trees, all more than 1 NNI away from each other. Trees correspond to x-NNI example in Serra Silva and Wilkinson(202x).
Running the example should yield 2 islands, one with 2 trees and one with a single tree.
}
\examples{
data(twoNNItestTrees)
neighbourhoods <- nniNeighbourhoods(twoNNItestTrees)
islands <- nniIslands(neighbourhoods)
#write.islands(islands)
}
\keyword{datasets}
