% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.islands.R
\name{write.islands}
\alias{write.islands}
\title{write.islands Function}
\usage{
write.islands(islands)
}
\arguments{
\item{islands}{list of objects of class "multiphylo"}
}
\value{
None
}
\description{
This function writes the NNI islands identified by \code{\link{nniIslands}} to individual Newick format tree files.
}
\examples{
data(testTrees)
neighbourhoods <- nniNeighbourhoods(testTrees)
islands <- nniIslands(neighbourhoods)
#write.islands(islands)

}
\references{
Maddison (1991) Syst Zool 40:315-328
(\href{https://doi.org/10.1093/sysbio/40.3.315}{OUP})
}
