% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xIslands.R
\name{xIslands}
\alias{xIslands}
\title{xIslands Function}
\usage{
xIslands(mat, tree, threshold, output = list(), verbose = TRUE)
}
\arguments{
\item{mat}{A distance matrix}

\item{tree}{An object of class "multiPhylo"}

\item{threshold}{A numeric value setting the x-RF island threshold}

\item{output}{'Dummy' parameter required for recursive function with changing number of inputs from first to subsequent rounds of the function. Please ALWAYS add 'output = list()' as a parameter.}

\item{verbose}{Prints the function's progress. Defaults to TRUE. If run in parallel set to FALSE.}
}
\value{
List of "multiPhylo" objects
}
\description{
This function recursively extracts the x-distance islands from a user supplied tree-to-tree distance matrix.
At this time the package might not accurately place unresolved trees into islands.
If you are expecting your tree distribution to be made up of large numbers of very small islands, set options(expressions=500000) before analysis.
}
\examples{
data(testTrees)
isles <- xRFislands(testTrees, threshold = 2, output = list())
#write.islands(islands)

}
\references{
Maddison, D. R. (1991) \href{https://doi.org/10.1093/sysbio/40.3.315}{The discovery and importance of multiple islands of most-parsimonious trees}. \emph{Syst. Zool.}, 40:315-328
}
