% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xRFislands.R
\name{xRFislands}
\alias{xRFislands}
\title{xRFislands Function}
\usage{
xRFislands(
  tree,
  threshold,
  output = list(),
  verbose = TRUE,
  checkUnique = TRUE
)
}
\arguments{
\item{tree}{An object of class "multiPhylo"}

\item{threshold}{A numeric value setting the x-RF island threshold}

\item{output}{'Dummy' parameter required for recursive function with changing number of inputs from first to subsequent rounds of the function. Please ALWAYS add 'output = list()' as a parameter.}

\item{verbose}{Prints the function's progress. Defaults to TRUE. If run in parallel set to FALSE.}

\item{checkUnique}{Parameter to check tree distribution is made up of unique trees. Defaults to TRUE for first round of recursion.}
}
\value{
List of "multiPhylo" objects
}
\description{
This function recursively computes the Robinson-Foulds distance matrix for a tree distribution and extracts the x-RF islands.
At this time the package might not accurately place unresolved trees into islands.
}
\examples{
data(testTrees)
isles <- xRFislands(testTrees, threshold = 2, output = list())
#write.islands(islands)

}
\references{
Maddison, D. R. (1991) \href{https://doi.org/10.1093/sysbio/40.3.315}{The discovery and importance of multiple islands of most-parsimonious trees}. \emph{Syst. Zool.}, 40:315-328
}
