% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xSDislands.R
\name{xSDislands}
\alias{xSDislands}
\title{xSDislands Function}
\usage{
xSDislands(tree, threshold, output = list())
}
\arguments{
\item{tree}{An object of class "multiPhylo"}

\item{threshold}{A numeric value setting the x-SD island threshold}

\item{output}{'Dummy' parameter required for recursive function with changing number of inputs from first to subsequent rounds of the function. Please ALWAYS add 'output = list()' as a parameter.}
}
\value{
List of "multiPhylo" objects
}
\description{
This function recursively computes the Robinson-Foulds/Symmetric Difference on Full Splits distance matrix for a tree distribution and extracts the x-SD islands.
The
At this time the package might not accurately place unresolved trees into islands.
}
\examples{
data(testTrees)
isles <- xSDislands(testTrees, threshold = 2, output = list())
#write.islands(islands)

}
\references{
Maddison, D. R. (1991) \href{https://doi.org/10.1093/sysbio/40.3.315}{The discovery and importance of multiple islands of most-parsimonious trees}. \emph{Syst. Zool.}, 40:315-328
}
