% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xSDneighbourhoods.R
\name{xSDneighbourhoods}
\alias{xSDneighbourhoods}
\title{xSDneighbourhoods Function}
\usage{
xSDneighbourhoods(tree, verbose = TRUE, threshold)
}
\arguments{
\item{tree}{an object of class "multiPhylo"}

\item{verbose}{Prints the function's progress. Defaults to TRUE. If run in parallel set to FALSE.}

\item{threshold}{A numeric value setting the x-SD island threshold}
}
\value{
List of "multiPhylo" objects
}
\description{
This function extracts the x-SD neighbourhoods from a multiPhylo object, it generates a list of multiPhylo objects containing all trees shared between each tree's x-SD neighbourhood and the multiPhylo object.
At this time the package might not accurately place unresolved trees into islands.
}
\examples{
data(testTrees)
neighbourhoods <- xSDneighbourhoods(testTrees)
islands <- neighbourhoodMerger(neighbourhoods)
#write.islands(islands)

}
\references{
Maddison, D. R. (1991) \href{https://doi.org/10.1093/sysbio/40.3.315}{The discovery and importance of multiple islands of most-parsimonious trees}. \emph{Syst. Zool.}, 40:315-328
}
