% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nniNeighbourhoods.R
\name{nniNeighbourhoods}
\alias{nniNeighbourhoods}
\title{nniNeighbourhoods Function}
\usage{
nniNeighbourhoods(tree, verbose = TRUE)
}
\arguments{
\item{tree}{an object of class "multiPhylo"}

\item{verbose}{Prints the function's progress. Defaults to TRUE. If run in parallel set to FALSE.}
}
\value{
List of "multiPhylo" objects
}
\description{
This function extracts the 1-NNI neighbourhoods from a multiPhylo object, it generates a list of multiPhylo objects containing all trees shared between each tree's NNI neighbourhood and the multiPhylo object.
At this time the package cannot place unresolved trees into islands.
}
\examples{
data(testTrees)
neighbourhoods <- nniNeighbourhoods(testTrees)
islands <- neighbourhoodMerger(neighbourhoods)
#write.islands(islands)

}
\references{
Maddison, D. R. (1991) \href{https://doi.org/10.1093/sysbio/40.3.315}{The discovery and importance of multiple islands of most-parsimonious trees}. \emph{Syst. Zool.}, 40:315-328
}
