% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testTrees-data.R
\docType{data}
\name{testTrees}
\alias{testTrees}
\title{Example trees for islandNeighbours package}
\format{
An object of class "multiPhylo".
}
\source{
\href{https://github.com/anaserrasilva/MajorityRuleAndTreeIslands}{https://github.com/anaserrasilva/MajorityRuleAndTreeIslands}
}
\usage{
data(testTrees)
}
\description{
Selection of ten phylogenetic trees from the set of most parsimonious trees recovered from Pardo et al.'s (2017) datamatrix.
Running the example should yield 2 islands, each with 5 trees.
}
\examples{
data(testTrees)
neighbourhoods <- nniNeighbourhoods(testTrees)
islands <- nniIslands(neighbourhoods)
#write.islands(islands)
}
\references{
Pardo et al. (2017) \href{https://pubmed.ncbi.nlm.nih.gov/28630337/}{Stem caecilian from the Triassic of Colorado sheds light on the origins of Lissamphibia}. \emph{PNAS}, 114:5389-5395

Maddison, D. R. (1991) \href{https://doi.org/10.1093/sysbio/40.3.315}{The discovery and importance of multiple islands of most-parsimonious trees}. \emph{Syst. Zool.}, 40:315-328
}
\keyword{datasets}
