% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.islands.R
\name{write.islands}
\alias{write.islands}
\title{write.islands Function}
\usage{
write.islands(islands)
}
\arguments{
\item{islands}{list of objects of class "multiPhylo"}
}
\value{
None
}
\description{
This function writes the NNI islands identified by \code{\link{nniIslands}} to individual Newick format tree files.
}
\examples{
data(testTrees)
neighbourhoods <- nniNeighbourhoods(testTrees)
islands <- nniIslands(neighbourhoods)
#write.islands(islands)

}
\references{
Maddison, D. R. (1991) \href{https://doi.org/10.1093/sysbio/40.3.315}{The discovery and importance of multiple islands of most-parsimonious trees}. \emph{Syst. Zool.}, 40:315-328
}
